add coordinates for small areas, change variables to snake case
