+++ fields.json
@@ -1,28 +1,34 @@
 {
     "3v3": {
-        "fieldSize": [1.500, 1.300],
-        "sAreaSize": [0.150, 0.700],
-        "quadRef": {
+        "field_size": [1.500, 1.300],
+        "small_area": {
+            "offensive": [1.300, 0.300, 0.150, 0.700],
+            "defensive": [0, 0.300, 0.150, 0.700]
+        },
+        "quad_ref": {
             "q1": [1.125, 1.050],
             "q2": [0.375, 1.050],
             "q3": [0.375, 0.250],
             "q4": [1.125, 0.250]
         },
-        "freeKick": {
+        "free_kick": {
             "left": [0.375, 0.650],
             "right": [1.125, 0.650]
         }
     },
     "5v5": {
-        "fieldSize": [2.200, 1.800],
-        "sAreaSize": [0.150, 0.500],
-        "quadRef": {
+        "field_size": [2.200, 1.800],
+        "small_area": {
+            "offensive": [2.050, 0.650, 0.150, 0.500],
+            "defensive": [0, 0.650, 0.150, 0.500]
+        },
+        "quad_ref": {
             "q1": [1.650, 1.500],
             "q2": [0.550, 1.500],
             "q3": [0.550, 0.300],
             "q4": [1.650, 0.300]
         },
-        "freeKick": {
+        "free-kick": {
             "left": [0.375, 0.900],
             "right": [1.825, 0.900]
         }
